apps/cipas-helper-services/code-clone-detection/requirements.txt
# Core dependencies
pyyaml>=6.0
pandas>=2.0.0
pyarrow>=12.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional dependencies
# tree-sitter>=0.20.0  # For AST parsing
# numpy>=1.24.0  # For numerical operations
# scikit-learn>=1.3.0  # For similarity metrics
